Move estrutura de navegação para screen_systems
