--- src/draw_states/mod.rs
@@ -1,1 +1,0 @@
-pub mod navigation;
